terror: remove porting
